Unite Contexts and add more user values
