--- Contexts/UserContextFactory.cs
@@ -1,22 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Design;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Contexts
-{
-    public class UserContextFactory : IDesignTimeDbContextFactory<UserContext>
-    {
-        public UserContext CreateDbContext(string[] args)
-        {
-            var optionsBuilder = new DbContextOptionsBuilder<UserContext>();
-            optionsBuilder.UseSqlite(
-                "Data Source=Database.db");
-
-            return new UserContext(optionsBuilder.Options);
-        }
-    }
-}
