+++ Models/UserModel.cs
@@ -5,7 +5,8 @@
         public int id { get; set; }
         public string login { get; set; }
         public string name { get; set; }
-        public enum role { User, Admin }
+        public string role { get; set; }
         public string phone_number { get; set; }
+        public string password { get; set; }
     }
 }
